pyproject.toml
[tool.poetry]
name = "tech-giants-backend"
version = "0.1.0"
description = ""
authors = ["Svetovoyvv <svetovoyvv@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
black = "^24.8.0"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
psycopg2 = "^2.9.9"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
django-filter = "^24.3"
django-constance = "^4.1.2"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
